requirements.txt

pandas==2.2.2
SQLAlchemy==2.0.29

psycopg2-binary==2.9.9

gspread==6.0.2
google-auth==2.29.0

python-dotenv==1.0.1